+++ packages/base/Fade/src/Fade/story/Default.example.tsx
@@ -1,6 +1,5 @@
 import { Container, Button } from '@toptal/picasso'
 import { Fade } from '@toptal/picasso-fade'
-import { SPACING_4 } from '@toptal/picasso-utils'
 import React, { useState } from 'react'
 
 const Example = () => {
@@ -9,10 +8,12 @@ const Example = () => {
   const handleOnClick = () => setFaded(prevFaded => !prevFaded)
 
   return (
-    <Container bottom={SPACING_4}>
-      <Button onClick={handleOnClick}>Toggle Fade</Button>
-      <Fade in={faded} timeout={3000}>
-        <div className='bg-red-500 p-4'>Fade in</div>
+    <Container>
+      <Button onClick={handleOnClick} className='mb-4'>
+        Toggle Fade
+      </Button>
+      <Fade in={faded} timeout={350}>
+        <div className='bg-gray-100 p-4 rounded-md'>Fade in content</div>
       </Fade>
     </Container>
   )