+++ packages/base/Grid/src/GridContext/index.ts
@@ -1,0 +1,1 @@
+export { default as GridContext } from './GridContext'