+++ packages/base/Pagination/src/Pagination/Pagination.tsx
@@ -75,7 +75,7 @@ export const Pagination = forwardRef<HTMLDivElement, Props>(function Pagination(
     return (
       <PaginationButton
         key={index}
-        className={classes.button}
+        className='[&+&]:ml-2'
         page={Number(page)}
         activePage={activePage}
         disabled={disabled}
@@ -87,7 +87,7 @@ export const Pagination = forwardRef<HTMLDivElement, Props>(function Pagination(
   return (
     <Container {...rest} ref={ref} flex inline alignItems='center'>
       <Button
-        className={classes.button}
+        className='[&+&]:ml-2'
         disabled={isFirstActive || disabled}
         onClick={handlePrevClick}
         variant='secondary'
@@ -99,7 +99,7 @@ export const Pagination = forwardRef<HTMLDivElement, Props>(function Pagination(
       {variant === 'compact' ? null : pageButtons}
 
       <Button
-        className={classes.button}
+        className='[&+&]:ml-2'
         disabled={isLastActive || disabled}
         onClick={handleNextClick}
         variant='secondary'