+++ packages/shared/src/styles/index.ts
@@ -50,6 +50,4 @@ export const outline = (baseColor: string, width = 3) => ({
 
 export const remToNumber = (value: string) => Number.parseFloat(value)
 
-export { default as addClassesToChildren } from './add-classes-to-children'
-
 export * from './types'