+++ packages/base/Popper/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @toptal/picasso-popper
 
+## 1.0.3
+
+### Patch Changes
+
+- Updated dependencies [[`d599529`](https://github.com/toptal/picasso/commit/d599529bcb283c367b63c612fee81394e66c9740)]:
+  - @toptal/picasso-shared@15.0.0
+  - @toptal/picasso-utils@1.0.3
+
 ## 1.0.2
 
 ### Patch Changes