+++ packages/base/Test-Utils/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @toptal/picasso-test-utils
 
+## 1.1.1
+
+### Patch Changes
+
+- Updated dependencies [[`d599529`](https://github.com/toptal/picasso/commit/d599529bcb283c367b63c612fee81394e66c9740)]:
+  - @toptal/picasso-shared@15.0.0
+
 ## 1.1.0
 
 ### Minor Changes