+++ packages/picasso-pictograms/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 4.0.0
+
+### Patch Changes
+
+- Updated dependencies [[`d599529`](https://github.com/toptal/picasso/commit/d599529bcb283c367b63c612fee81394e66c9740)]:
+  - @toptal/picasso-shared@15.0.0
+  - @toptal/picasso-utils@1.0.3
+
 ## 3.0.1
 
 ### Patch Changes